backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
supabase==2.4.6
reportlab==4.2.0
requests==2.31.0
pillow==10.3.0
python-dotenv==1.0.1
gunicorn==22.0.0
